Fix doctor availability check in AppointmentService

+++ SymptomCheck-backend/doctorservice/src/main/java/com/symptomcheck/doctorservice/repositories/DoctorProfileRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DoctorProfileRepository extends JpaRepository<DoctorProfile, Integer> {
+public interface DoctorProfileRepository extends JpaRepository<DoctorProfile, Long> {
 }
